Move USBTransferType from USBPipe.cs to USB.cs
