+++ WinUSBNet/USBPipe.cs
@@ -13,24 +13,6 @@ using Nefarius.Drivers.WinUSB.API;
 namespace Nefarius.Drivers.WinUSB;
 
 /// <summary>
-///     USB transfer type enumeration
-/// </summary>
-public enum USBTransferType
-{
-    /// <summary>The pipe is a control transfer pipe</summary>
-    Control = USBD_PIPE_TYPE.UsbdPipeTypeControl,
-
-    /// <summary>The pipe is an isochronous transfer pipe</summary>
-    Isochronous = USBD_PIPE_TYPE.UsbdPipeTypeIsochronous,
-
-    /// <summary>The pipe is a bulk transfer pipe</summary>
-    Bulk = USBD_PIPE_TYPE.UsbdPipeTypeBulk,
-
-    /// <summary>The pipe is an interrupt transfer pipe</summary>
-    Interrupt = USBD_PIPE_TYPE.UsbdPipeTypeInterrupt,
-}
-
-/// <summary>
 ///     UsbPipe represents a single pipe on a WinUSB device. A pipe is connected
 ///     to a certain endpoint on the device and has a fixed direction (IN or OUT)
 /// </summary>
